add python 3.6 unit test job
See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

This also modifies a unit test to not attempt to pass the re.LOCALE flag
when the regex is a string, which is prohibited in Python 3.6. It would
be possible to use a raw bytes object as the regex instead, but we're
testing the repr() here and it has a different representation between
Python 2 and Python 3. Just use the re.IGNORECASE flag instead for the
purposes of the test.

Change-Id: I666d49d0a29963bcaef6ca69be34dc4869f6db6f
Co-Authored-By: Zane Bitter <zbitter@redhat.com>
Story: #2002586
Task: #24322