Remove flake8 from the test-requirements
flake8 is a linter so we don't need to pull in for the unit
tests or other tox environments.

Also update to the newer flake8 2.6.x which is the version
chosen by current hacking release (which causes no fallout so should be fine).
Also this is the first version that is compatible being run
under python 3.6/7.

Change-Id: I54f99e7e29c7b3727db98bdaf8bc168da3890a1d